Pass on props for title and author and category
