+++ src/Pages/Categories.js
@@ -1,0 +1,10 @@
+const Categories = () => (
+  <div>
+    <ul>
+      <h3>Categories</h3>
+      <li>All</li>
+    </ul>
+  </div>
+);
+
+export default Categories;
